Rename topkTokens dir
